content of geometry.utilities got split into morphology and transformations modules
