+++ docs/source/eolearn.geometry.morphology.rst
@@ -1,0 +1,7 @@
+eolearn.geometry.morphology
+===========================
+
+.. automodule:: eolearn.geometry.morphology
+    :members:
+    :undoc-members:
+    :show-inheritance:
