+++ docs/source/eolearn.geometry.rst
@@ -10,6 +10,7 @@ Submodules:
 
 .. toctree::
 
+   eolearn.geometry.morphology
    eolearn.geometry.sampling
    eolearn.geometry.superpixel
-   eolearn.geometry.utilities
+   eolearn.geometry.transformations
