--- docs/source/eolearn.geometry.utilities.rst
@@ -1,7 +1,0 @@
-eolearn.geometry.utilities
-==========================
-
-.. automodule:: eolearn.geometry.utilities
-    :members:
-    :undoc-members:
-    :show-inheritance:
